+++ apm-sniffer/apm-sdk-plugin/graphql-plugin/graphql-12.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/graphql/v12/define/GraphqlInstrumentation.java
@@ -64,6 +64,9 @@ public class GraphqlInstrumentation extends ClassInstanceMethodsEnhancePluginDef
 
     @Override
     protected String[] witnessClasses() {
-        return new String[] {"graphql.cachecontrol.CacheControl"};
+        return new String[] {
+                "graphql.cachecontrol.CacheControl",
+                "graphql.execution.ExecutionPath"
+        };
     }
 }