+++ CHANGES.md
@@ -13,6 +13,7 @@ Release Notes.
 * Support Dbcp2 Connection pool metrics collecting.
 * Ignore the synthetic constructor created by the agent in the Spring patch plugin.
 * Add witness class for vertx-core-3.x plugin.
+* Add witness class for graphql plugin.
 
 #### Documentation